requirements.txt
https://www.djangoproject.com/download/1.8a1/tarball/
psycopg2
django-redis-sessions
redis
hiredis
arrow
requests
beanstalkc
pyyaml
sevencow
